changes data from app and refactoring
