+++ app/src/main/res/drawable/ic_cyclone.xml
