+++ app/src/main/res/drawable/ic_hive.xml
