+++ app/src/main/res/drawable/ic_smile.xml
