+++ app/src/main/res/values/colors.xml
@@ -5,7 +5,7 @@
     <color name="colorPrimaryVariant">#FF3700B3</color>
     <color name="colorSecondary">#FF03DAC5</color>
     <color name="colorSecondaryVariant">#FF018786</color>
-    <color name="colorOnSecondary">#FF000000</color>
+    <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
 
 </resources>
