+++ app/src/main/res/values/themes.xml
@@ -9,7 +9,7 @@
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/colorSecondary</item>
         <item name="colorSecondaryVariant">@color/colorSecondaryVariant</item>
-        <item name="colorOnSecondary">@color/colorOnSecondary</item>
+        <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
         <item name="android:statusBarColor">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
@@ -26,7 +26,7 @@
         <item name="android:button">@drawable/rb_selector</item>
         <item name="android:padding">8dp</item>
         <item name="android:layout_margin">8dp</item>
-        <item name="android:textColor">@color/text_color_selector</item>
+        <item name="android:textColor">@color/text_selector</item>
         <item name="android:textSize">8sp</item>
     </style>
