+++ ss-smart-filter/src/main/java/com/ss/smartfilterlib/data/SingleSelectionParams.kt
@@ -3,21 +3,21 @@ package com.ss.smartfilterlib.data
 import android.view.ViewGroup
 import androidx.annotation.ColorRes
 import androidx.annotation.DrawableRes
-import androidx.appcompat.R
+import com.ss.smartfilterlib.R
+
 import com.ss.smartfilterlib.utils.SingleChipType
-import com.ss.smartfilterlib.utils.SingleGroupSubType
+
 
 /**
  * created by Mala Ruparel ON 08/05/24
  */
 data class SingleSelectionParams(
     val rootView: ViewGroup,
-    val singleGroupSubType: SingleGroupSubType,
     val chipType: SingleChipType,
     val orientation: Int,
     val mData: ArrayList<Data>,
     val onItemSelected: (Data) -> Unit,
-    @DrawableRes val bgSelector: Int = R.drawable.abc_btn_radio_material,
-    @ColorRes val textSelector: Int = android.R.color.black
+    @DrawableRes val bgSelector: Int = R.color.chip_bg_selector,
+    @ColorRes val textSelector: Int = R.color.chip_text_selector
 
 )
