+++ ss-smart-filter/src/main/java/com/ss/smartfilterlib/singleselection/SingleSelectionMultiLineRadioButton.kt
@@ -197,11 +197,11 @@ class SingleSelectionMultiLineRadioButton @JvmOverloads constructor(context: Con
     }
 
     private fun setDefaultDrawable(): Drawable? {
-        return ContextCompat.getDrawable(context, R.drawable.multiline_default)
+        return ContextCompat.getDrawable(context, R.drawable.multiline_bg_selector)
     }
 
     private fun setDefaultTextColor(): ColorStateList? {
-        return ContextCompat.getColorStateList(context, R.color.colorOnSecondary)
+        return ContextCompat.getColorStateList(context, R.color.multiline_text_selector)
     }
