+++ ss-smart-filter/src/main/java/com/ss/smartfilterlib/utils/Constant.kt
@@ -9,4 +9,5 @@ object Constant {
     const val PRIMARY_TEXT_COLOR = Color.BLACK
     const val DEFAULT_SPAN_COUNT = 3
     const val DEFAULT_SPACING = 4
+
 }
