+++ ss-smart-filter/src/main/res/color/chip_bg_selector.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:color="@color/purple_200"
+        android:color="@color/colorPrimary"
         android:state_checked="true"
         />
