+++ ss-smart-filter/src/main/res/color/multiline_text_selector.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_pressed="true" android:color="@color/purple_200"/>
-    <item android:state_checked="true" android:color="@color/colorOnPrimary"/>
-    <item android:color="@color/colorOnSecondary"/>
+    <item android:state_pressed="true" android:color="@color/colorPrimary"/>
+    <item android:state_checked="true" android:color="@color/white"/>
+    <item android:color="@color/black"/>
 </selector>
