--- ss-smart-filter/src/main/res/color/text_color_selector.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_checked="false" android:color="@color/colorOnSecondary"/>
-    <item android:state_checked="true" android:color="@color/purple_200"/>
-    <item android:color="@color/colorOnSecondary"/>
-
-</selector>
