+++ ss-smart-filter/src/main/res/color/text_selector.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_pressed="true" android:color="@color/colorPrimary"/>
-    <item android:state_checked="true" android:color="@color/white"/>
+    <item android:state_checked="false" android:color="@color/black"/>
+    <item android:state_checked="true" android:color="@color/colorPrimary"/>
     <item android:color="@color/black"/>
+
 </selector>
