+++ ss-smart-filter/src/main/res/drawable/ic_star_filled.xml
@@ -1,7 +1,7 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="24dp"
     android:height="24dp"
-    android:tint="@color/purple_200"
+    android:tint="@color/colorPrimary"
     android:viewportWidth="24"
     android:viewportHeight="24">
