+++ ss-smart-filter/src/main/res/drawable/multiline_default.xml
@@ -6,7 +6,7 @@
 
     <stroke
         android:width="1dp"
-        android:color="@color/colorPrimaryVariant" />
+        android:color="@color/colorPrimary" />
 
     <corners android:radius="8dp" />
