+++ ss-smart-filter/src/main/res/drawable/multiline_selected.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="@color/colorPrimaryVariant" />
+    <solid android:color="@color/colorPrimary" />
     <corners android:radius="8dp" />
 
 </shape>
