+++ ss-smart-filter/src/main/res/drawable/row_item_pressed.xml
@@ -6,7 +6,7 @@
         android:left="16dp"
         android:right="16dp"
         android:top="0dp" />
-    <solid android:color="@color/purple_200" />
+    <solid android:color="@color/colorPrimary" />
     <corners android:radius="4dp" />
 
 </shape>
