+++ ss-smart-filter/src/main/res/layout/row_item.xml
@@ -35,7 +35,7 @@
             android:ellipsize="end"
             android:includeFontPadding="false"
             android:maxLines="1"
-            android:textColor="@color/colorOnSecondary"
+            android:textColor="@color/black"
             android:textSize="14sp"
             tools:text="Title" />
 
@@ -52,7 +52,7 @@
             android:ellipsize="end"
             android:gravity="start"
             android:maxLines="2"
-            android:textColor="@color/colorOnSecondary"
+            android:textColor="@color/black"
             android:textSize="@dimen/_11ssp"
             tools:text="Description" />
