+++ ss-smart-filter/src/main/res/values/colors.xml
@@ -6,8 +6,8 @@
     <color name="colorPrimaryVariant">#FF3700B3</color>
     <color name="colorSecondary">#FF03DAC5</color>
     <color name="colorSecondaryVariant">#FF018786</color>
-    <color name="colorOnSecondary">#FF000000</color>
-    <color name="colorOnPrimary">#ffffff</color>
+    <color name="black">#FF000000</color>
+    <color name="white">#ffffff</color>
     <color name="grey">#D3D3D3</color>
     <color name="grey_200">#FFEEEEEE</color>
     <color name="grey_500">#FF9E9E9E</color>
