+++ ss-smart-filter/src/main/res/values/themes.xml
@@ -10,7 +10,7 @@
     </style>
 
     <style name="itemDescStyle">
-        <item name="android:textColor">@color/colorOnSecondary</item>
+        <item name="android:textColor">@color/black</item>
         <item name="android:textSize">14sp</item>
         <item name="android:layout_centerHorizontal">true</item>
         <item name="android:paddingTop">0dp</item>
@@ -18,9 +18,9 @@
 
     <style name="EntryChipStyle" parent="Widget.MaterialComponents.Chip.Entry">
         <item name="chipBackgroundColor">@color/purple_200</item>
-        <item name="chipStrokeColor">@color/colorOnSecondary</item>
-        <item name="android:textColor">@color/colorOnSecondary</item>
-        <item name="closeIconTint">@color/colorOnSecondary</item>
+        <item name="chipStrokeColor">@color/black</item>
+        <item name="android:textColor">@color/black</item>
+        <item name="closeIconTint">@color/black</item>
     </style>
     <style name="FilterChipStyle" parent="Widget.MaterialComponents.Chip.Filter"/>
